測試分離路由booking.py是否能正常運作
